+++ doc/d-practical-exercises/src/foot_steps.py
@@ -1,21 +1,22 @@
-class FootSteps(object):
+class FootSteps:
     """
     The class stores three functions of time: left, right and flying_foot.
     Each function is piecewise constant. For each function, the user can ask
     what is the value of this function at time t.
 
-    The storage is composed of three lists for left, right and flying_foot, and a list for time.
+    The storage is composed of three lists for left, right and flying_foot, and a list
+    for time.
     The list of times stores the time intervals, i.e. each element of the list is
     the start of a time interval. The first element of the list is 0.
-    The value of the functions left, right, flying_foot one this time interval is stored at
-    the same position is their respective list (i.e. value of left on interval
+    The value of the functions left, right, flying_foot one this time interval is stored
+    at the same position is their respective list (i.e. value of left on interval
     [time[i], time[i+1]] is stored in left[i].
 
     The 4 lists are set up using function add_phase().
-    The values of functions left, right, flying_foot can be accessed through the function
-    get_phase_type(t), get_left_position(t), get_right_position(t).
-    phase_type are 'left' (meaning left foot is flying, right foot is fixed), 'right' (ie the opposite)
-    or 'none' (meaning no foot is flying, both are fixed on the ground).
+    The values of functions left, right, flying_foot can be accessed through the
+    function get_phase_type(t), get_left_position(t), get_right_position(t).  phase_type
+    are 'left' (meaning left foot is flying, right foot is fixed), 'right' (ie the
+    opposite) or 'none' (meaning no foot is flying, both are fixed on the ground).
 
     Additionnally, functions get_left_next_position(t),
     get_right_next_position(t) can be used to get the next position of the
@@ -39,10 +40,10 @@ class FootSteps(object):
 
     def add_phase(self, duration, foot, position=None):
         """
-        Add a phase lasting <duration> where the flyhing foot <foot> (either 'left' or 'right')
-        moves to <position> (being a vector or a SE3 placement).
-        Alternatively, <foot> might be set to 'none' (i.e double support). In that case, <position>
-        is not specified (or is set to None, default).
+        Add a phase lasting <duration> where the flyhing foot <foot> (either 'left' or
+        'right') moves to <position> (being a vector or a SE3 placement).
+        Alternatively, <foot> might be set to 'none' (i.e double support). In that case,
+        <position> is not specified (or is set to None, default).
         """
         assert foot == "left" or foot == "right" or foot == "none"
         self.time.append(self.time[-1] + duration)
@@ -55,7 +56,9 @@ class FootSteps(object):
             self.right[-1] = position
 
     def get_index_from_time(self, t):
-        """Return the index i of the interval containing t, i.e. t in time[i], time[i+1]"""
+        """
+        Return the index i of the interval containing t, i.e. t in time[i], time[i+1]
+        """
         if t > self.time[-1]:
             return len(self.time) - 1
         return next(i for i, ti in enumerate(self.time) if ti > t) - 1