+++ unittest/python/bindings_liegroups.py
@@ -2,7 +2,6 @@ import unittest
 
 import numpy as np
 import pinocchio as pin
-
 from test_case import PinocchioTestCase as TestCase