+++ unittest/python/bindings_model.py
@@ -1,7 +1,7 @@
 import unittest
+
 import pinocchio as pin
 from pinocchio.utils import np, zero
-
 from test_case import PinocchioTestCase as TestCase