+++ unittest/python/bindings_rnea.py
@@ -1,8 +1,8 @@
 import unittest
-from test_case import PinocchioTestCase as TestCase
 
-import pinocchio as pin
 import numpy as np
+import pinocchio as pin
+from test_case import PinocchioTestCase as TestCase
 
 
 class TestRNEA(TestCase):