+++ unittest/python/utils.py
@@ -4,7 +4,6 @@ from math import sqrt
 import numpy as np
 import pinocchio as pin
 from pinocchio.utils import isapprox
-
 from test_case import PinocchioTestCase as TestCase